+++ TICoreDataSync/02 Primary Classes/TICDSDocumentSyncManager.m
@@ -2287,7 +2287,7 @@
     self.backgroundTaskID = [[UIApplication sharedApplication] beginBackgroundTaskWithExpirationHandler:^{
                                  [self endBackgroundTask];
                              }];
-    TICDSLog(TICDSLogVerbosityEveryStep, @"Doc Sync Manager (%@), Task ID (%i) is begining.", [self class], self.backgroundTaskID);
+    TICDSLog(TICDSLogVerbosityEveryStep, @"Doc Sync Manager (%@), Task ID (%lu) is begining.", [self class], (unsigned long)self.backgroundTaskID);
 #endif
 }
 
@@ -2300,13 +2300,13 @@
 
     switch ([[UIApplication sharedApplication] applicationState]) {
         case UIApplicationStateActive:  {
-            TICDSLog(TICDSLogVerbosityEveryStep, @"Doc Sync Manager (%@), Task ID (%i) is ending while app state is Active", [self class], self.backgroundTaskID);
+            TICDSLog(TICDSLogVerbosityEveryStep, @"Doc Sync Manager (%@), Task ID (%lu) is ending while app state is Active", [self class], (unsigned long)self.backgroundTaskID);
         }   break;
         case UIApplicationStateInactive:  {
-            TICDSLog(TICDSLogVerbosityEveryStep, @"Doc Sync Manager (%@), Task ID (%i) is ending while app state is Inactive", [self class], self.backgroundTaskID);
+            TICDSLog(TICDSLogVerbosityEveryStep, @"Doc Sync Manager (%@), Task ID (%lu) is ending while app state is Inactive", [self class], (unsigned long)self.backgroundTaskID);
         }   break;
         case UIApplicationStateBackground:  {
-            TICDSLog(TICDSLogVerbosityEveryStep, @"Doc Sync Manager (%@), Task ID (%i) is ending while app state is Background with %.0f seconds remaining", [self class], self.backgroundTaskID, [[UIApplication sharedApplication] backgroundTimeRemaining]);
+            TICDSLog(TICDSLogVerbosityEveryStep, @"Doc Sync Manager (%@), Task ID (%lu) is ending while app state is Background with %.0f seconds remaining", [self class], (unsigned long)self.backgroundTaskID, [[UIApplication sharedApplication] backgroundTimeRemaining]);
         }   break;
         default:
             break;
@@ -2323,21 +2323,21 @@
     @synchronized(self) {
         for (TICDSOperation *op in [self.registrationQueue operations]) {
             if (!op.shouldContinueProcessingInBackgroundState) {
-                TICDSLog(TICDSLogVerbosityEveryStep,@"Doc Sync Manager is cancelling operation %@ (id:%i) because app has gone to background state.",[self class],self.backgroundTaskID);
+                TICDSLog(TICDSLogVerbosityEveryStep,@"Doc Sync Manager is cancelling operation %@ (id:%lu) because app has gone to background state.",[self class],(unsigned long)self.backgroundTaskID);
                 [op cancel];
             }
         }
         
         for (TICDSOperation *op in [self.synchronizationQueue operations]) {
             if (!op.shouldContinueProcessingInBackgroundState) {
-                TICDSLog(TICDSLogVerbosityEveryStep,@"Doc Sync Manager is cancelling operation %@ (id:%i) because app has gone to background state.",[self class],self.backgroundTaskID);
+                TICDSLog(TICDSLogVerbosityEveryStep,@"Doc Sync Manager is cancelling operation %@ (id:%lu) because app has gone to background state.",[self class],(unsigned long)self.backgroundTaskID);
                 [op cancel];
             }
         }
 
         for (TICDSOperation *op in [self.otherTasksQueue operations]) {
             if (!op.shouldContinueProcessingInBackgroundState) {
-                TICDSLog(TICDSLogVerbosityEveryStep,@"Doc Sync Manager is cancelling operation %@ (id:%i) because app has gone to background state.",[self class],self.backgroundTaskID);
+                TICDSLog(TICDSLogVerbosityEveryStep,@"Doc Sync Manager is cancelling operation %@ (id:%lu) because app has gone to background state.",[self class],(unsigned long)self.backgroundTaskID);
                 [op cancel];
             }
         }
